Start consolidating SAML rules
