+++ rules/SAML-configuration-mapping.json
