--- rules/SAML-thinksmart.json
@@ -1,4 +1,0 @@
-{
-    "enabled": true,
-    "order": 700
-}
